+++ app/Services/TagService.php
@@ -2,6 +2,7 @@
 
 namespace App\Services;
 
+use App\Http\Responses\Response;
 use App\Models\Category;
 use App\Models\Course;
 use App\Models\CourseTag;
@@ -9,6 +10,8 @@ use App\Models\Tag;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
 use mysql_xdevapi\Exception;
+use function Laravel\Prompts\table;
+use function PHPUnit\Framework\isNull;
 
 class TagService
 {
@@ -39,8 +42,8 @@ class TagService
     public function createTag(string $tagName, $categoryId): Tag
     {
         try {
-            $category = Category::findOrFail($categoryId);
-
+            $category = Category::find($categoryId);
+            if (!$category) throw new \Exception(__('messages.category_not_found'));
             $tag = Tag::create([
                 'name' => $tagName,
                 'category_id' => $category->id,
@@ -49,40 +52,46 @@ class TagService
             return $tag;
         } catch (\Exception $e) {
             DB::rollBack();
-            throw $e;
+            throw new \Exception(__('messages.failed_to_create_tag', ['error' => $e->getMessage()]));
         }
     }
 
     public function addTagsToCourse($course_id, $tag_ids)
     {
         try {
-            $course = Course::findOrFail($course_id);
+            $course = Course::find($course_id);
+            if (!$course) throw new \Exception(__('messages.course_not_found'));
             $course->tags()->attach($tag_ids);
             return $course->tags;
         } catch (\Exception $e) {
-            throw new \Exception("Failed to add tags: " . $e->getMessage());
+            throw new \Exception(__('messages.failed_to_add_tags', ['error' => $e->getMessage()]));
         }
     }
 
     public function deleteTag($tagId): array
     {
         try {
-            $tag = Tag::findOrFail($tagId);
+            $tag = Tag::find($tagId);
+            if (!$tag) throw new \Exception(__('messages.tag_not_found'), 404);
             $tag->delete();
+
             DB::commit();
 
-            return ['message' => 'tag deleted successfully'];
+            return ['message' => __('messages.tag_deleted')];
         } catch (\Exception $e) {
             DB::rollBack();
-            throw $e;
+            throw new \Exception(__('messages.failed_to_add_tags', ['error' => $e->getMessage()]));
         }
     }
 
     public function updateTag($quizId, Request $request)
     {
         try {
-            $tag = Tag::findOrFail($quizId);
-            $category = Category::findOrFail($request['categoryId']);
+            $tag = Tag::find($quizId);
+            if (!$tag) throw new \Exception(__('messages.tag_not_found'), 404);
+            $category = Category::find($request['categoryId']);
+            if (!$category) throw new \Exception(__('messages.category_not_found'));
+
             if (isset($request['name'])) {
                 $tag->name = $request['name'];
                 $tag->save();
@@ -95,27 +104,23 @@ class TagService
 
         } catch (\Exception $e) {
             DB::rollBack();
-            throw $e;
+            throw new \Exception(__('messages.failed_to_update_tag', ['error' => $e->getMessage()]));
         }
     }
 
     public function getTagsByCategory($category_id)
     {
-        try {
-            $category = Category::with('tags')->findOrFail($category_id);
-            return $category->tags;
-        } catch (\Exception $e) {
-            throw new \Exception('Error fetching tags for category: ' . $e->getMessage());
-        }
+        $category = Category::with('tags')->find($category_id);
+        if (!$category) throw new \Exception(__('messages.category_not_found'), 404);
+        if (is_null($category->tags)) throw new \Exception(__('messages.no_tags'), 200);
+        return $category->tags;
     }
 
     public function getCourseTags($course_id)
     {
-        try {
-            $course = Course::with('tags')->findOrFail($course_id);
-            return $course->tags;
-        } catch (\Exception $e) {
-            throw new \Exception("Failed to fetch tags: " . $e->getMessage());
-        }
+        $course = Course::query()->firstWhere('id', $course_id);
+        if (!$course) throw new \Exception(__('messages.course_not_found'), 404);
+        if (!($course->tags)) throw new \Exception(__('messages.no_tags'), 404);
+        return $course->tags;
     }
 }
